requirements.txt
azure_storage==0.37.0
matplotlib==3.2.2
msrest==0.6.21
numpy==1.21.6
Pillow==9.1.0
requests==2.23.0
streamlit==1.9.0
tensorflow==2.8.0
tqdm==4.64.0
